Implement /stopinfo, which gets data for train stations.
Add the /stopinfo method, which the JavaScript calls to get info about
a train station when the user clicks on it.

Include webmock and use it to test /bikedata and /stopinfo, with mock
HTTP requests that return hard-coded responses. Still need to parse the
XML returned by /bikedata to check its contents.

The Rails version now has all the functionality of the Sinatra version,
plus it's got tests!