SubwayLine, SubwayStation and SubwayStop models.
Migrations and classes for SubwayLine (like Red, Orange or Blue),
SubwayStation (like Alewife or Downtown Crossing.), and
SubwayStop (like RALE or RTDC or RODT, the last two of which are
the Red and Orange stops at Downtown Crossing) models.

Make db/seeds.rb read the text file containing the data and generate
the SubwayStation and SubwayStop objects.