When serving /, get subway station info from the database, not the text file.
Use the SubwayStation, SubwayStop and SubwayLine classes to pull all the info
for subway stations. Previously this information would be read from
app/datafiles/stops.csv every time the page was requested. Make sure to load
the SubwayStop and SubwayLine data in the initial fetch, so there are only 3
selects run, rather than a number proportional to the number of stops in the
system.